refactor the missing header
