+++ http-server/inc/serverModelSimple.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "securityModel.h"
+#include "serverModel.h"
 #include <unordered_map>
 #include <thread>
